% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afegir_dataindex.R
\name{afegir_dataindex}
\alias{afegir_dataindex}
\title{Afegeix data index.}
\usage{
afegir_dataindex(dt_historic, bd.dindex = "20161231")
}
\arguments{
\item{dt_historic}{Base de dades medica estructura long amb idp:identificació pacient}

\item{bd.dindex}{Data Index ,que afagirem a la nostra base de dades}
}
\value{
Afageix la data index a la base de dades
}
\description{
Afegeix data index Dinamica o / Constant si no existeix
}
\examples{

variables_dtindex<-afegir_dataindex(dt_variables,bd.dindex=20081231)
variables_dtindex
}
