% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_analitiques.R
\name{agregar_analitiques}
\alias{agregar_analitiques}
\title{Agregar prescripcions.}
\usage{
agregar_analitiques(
  dt = ANALITIQUES,
  bd.dindex = "20161231",
  finestra.dies = c(-Inf, Inf),
  sufix = c(".valor", ".dies"),
  fun = "last",
  camp = "cod"
)
}
\arguments{
\item{dt}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{bd.dindex}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{finestra.dies}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{sufix}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{fun}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}
}
\value{
Taula agregada analitiques
}
\description{
Retorna tibble (data.table) amb els valors analitiques en temps previ en dies.
}
\examples{
u=rnorm(1000,100,12)

}
