% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_prescripcions.R
\name{agregar_prescripcions}
\alias{agregar_prescripcions}
\title{Agregar prescripcions.}
\usage{
agregar_prescripcions(
  dt = PRESCRIPCIONS,
  bd.dindex = 20161231,
  dt.agregadors = CATALEG,
  prefix = "FP.",
  finestra.dies = c(0, 0),
  camp_agregador = "agr",
  agregar_data = F,
  acumular = NULL,
  cataleg_mana = F
)
}
\arguments{
\item{dt}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{bd.dindex}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{dt.agregadors}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{prefix}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{finestra.dies}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{camp_agregador}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{agregar_data}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{acumular}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{cataleg_mana}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}
}
\value{
Taula agregada prescripcions
}
\description{
Retorna tibble (data.table) amb el temps de prescripció en una finestra o primera data per idp-dataindex / primera data
}
\examples{
u=rnorm(1000,100,12)

}
