% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_problemes.R
\name{agregar_problemes}
\alias{agregar_problemes}
\title{Agregar problemes de salut.}
\usage{
agregar_problemes(
  dt = PROBLEMES,
  bd.dindex = "20161231",
  dt.agregadors = CATALEG,
  finestra.dies = c(-Inf, 0),
  prefix = "DG.",
  camp_agregador = "agr",
  keep.code = F,
  cataleg_mana = F
)
}
\arguments{
\item{dt}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{bd.dindex}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{dt.agregadors}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{finestra.dies}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{prefix}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{camp_agregador}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{keep.code}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{cataleg_mana}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}
}
\value{
Taula agregada de problemes
}
\description{
Afegeix data-index Dinamica o / Constant si no existeix
}
\examples{
u=rnorm(1000,100,12)

}
