% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_solapaments_gaps.R
\name{agregar_solapaments_gaps}
\alias{agregar_solapaments_gaps}
\title{Agrega solapamentsamb gaps}
\usage{
agregar_solapaments_gaps(
  dt = dades,
  id = "idp",
  datainici = "data",
  datafinal = "datafi",
  gap = 5,
  sel = F
)
}
\arguments{
\item{dt}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{id}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{datainici}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{datafinal}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{gap}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{sel}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}
}
\value{
Retorna dades amb : id, datainici i datafi amb menys registres, havent eliminat solapaments i gaps (discontinuitat petita)
}
\description{
Historic de farmacs.Elimina solapaments i discontinuitats petites i retorna dades sense discontinuitats ni solapaments.
}
\examples{
u=rnorm(1000,100,12)

}
