% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_visites.R
\name{agregar_visites}
\alias{agregar_visites}
\title{Agregar visites.}
\usage{
agregar_visites(
  dt = VISITES,
  bd.dindex = 20161231,
  finestra.dies = c(-365, 0),
  N = "NA",
  data = "NA"
)
}
\arguments{
\item{dt}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{bd.dindex}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{finestra.dies}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{N}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{data}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}
}
\value{
Numero de visites en la finestra de temps
}
\description{
Envio la historic de visites i retorno numero de visites en la finestra de temps
}
\examples{
u=rnorm(1000,100,12)

}
