% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cataleg.R
\docType{data}
\name{cataleg}
\alias{cataleg}
\title{cataleg}
\format{
Una data frame amb 7163 files i 4 variables
\describe{
\item{domini}{domini (diagnostics farmacs )}
\item{cod}{cod (150.9 A10BA02 A10BB01 A10BD01 A10BD04 E11 I10 I25 J01DD07 Z94.9)}
\item{Descr}{Descr 150.9 A10BA02 A10BB01 A10BD01 A10BD04 E11 I10 I25 J01DD07 Z94.9)}
\item{agr}{agr (Antibioticos Biguanidas DM2 HTA INS.CARD
ISQ.CRONICA Sulfonilureas Tiazolidinadiones TRAST_SIST)}
}
}
\usage{
cataleg
}
\description{
cataleg
}
\keyword{datasets}
