% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_cataleg_test.R
\docType{data}
\name{dt_cataleg_test}
\alias{dt_cataleg_test}
\title{Cataleg}
\format{
Una data frame amb 9 files i 3 variables
\describe{
\item{domini}{domini (diagnostics farmacs )}
\item{cod}{cod (150.9 A10BA02 A10BB01 A10BD01 A10BD04 E11 I10 I25 J01DD07)}
\item{agr}{agr (Antibioticos Biguanidas DM2 HTA INS.CARD
ISQ.CRONICA Sulfonilureas Tiazolidinadiones)}
}
}
\usage{
dt_cataleg_test
}
\description{
Cataleg
}
\keyword{datasets}
