% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_diagnostics.R
\docType{data}
\name{dt_diagnostics}
\alias{dt_diagnostics}
\title{Prices of 50,000 round cut diamonds}
\format{
A data frame with 25 rows and 3 variables
\describe{
\item{idp}{idp (1--5)}
\item{cod}{cod (E11   I10   I25 I50.9 )}
\item{dat}{dat (20080115--20080215)}
}
}
\usage{
dt_diagnostics
}
\description{
Prices of 50,000 round cut diamonds
}
\keyword{datasets}
