% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_facturacio.R
\docType{data}
\name{dt_facturacio}
\alias{dt_facturacio}
\title{Bases de dades de medicaments facturats}
\format{
Una data frame amb 4421 files i 5 variables
\describe{
\item{idp}{idp (1--30)}
\item{cod}{codi (A10BA02 A10BB01 A10BD01 A10BD04 N05AL01 )}
\item{dat}{data (201001--202012)}
\item{env}{envasos (1--6)}
}
}
\usage{
dt_facturacio
}
\description{
Bases de dades de medicaments facturats
}
\keyword{datasets}
