% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_poblacio_test.R
\docType{data}
\name{dt_poblacio_test}
\alias{dt_poblacio_test}
\title{Bases de dades de la pobalcio}
\format{
Una data frame amb 30 files i 6 variables
\describe{
\item{idp}{idp (1--30)}
\item{sexe}{sexe (D H)}
\item{situacio}{situacio (A D T)}
\item{dnaix}{dnaix 19280901--19830901)}
\item{entrada}{entrada (20060101--20060101)}
\item{sortida}{sortida (20120228--20201231)}
}
}
\usage{
dt_poblacio_test
}
\description{
Bases de dades de la pobalcio
}
\keyword{datasets}
