% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_prescripcions.R
\docType{data}
\name{dt_prescripcions}
\alias{dt_prescripcions}
\title{Bases de dades de medicaments prescrits}
\format{
Una data frame amb 25 files i 4 variables
\describe{
\item{idp}{idp (1--5)}
\item{cod}{codi (A10BA02 A10BB01 A10BD01 A10BD04 J01DD07)}
\item{dat}{data (20080115--20080215)}
\item{dbaixa}{data_baixa (20080215--20080315)}
}
}
\usage{
dt_prescripcions
}
\description{
Bases de dades de medicaments prescrits
}
\keyword{datasets}
