% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_prescripcions_test.R
\docType{data}
\name{dt_prescripcions_test}
\alias{dt_prescripcions_test}
\title{Bases de dades de medicaments prescrits}
\format{
Una data frame amb 782 files i 5 variables
\describe{
\item{idp}{idp (1--30)}
\item{cod}{codi (A10BA02 A10BB01 A10BD01 A10BD04 J01DD07 N05AL01)}
\item{dat}{data (20051227--20201229)}
\item{dbaixa}{data_baixa (20100207--20220203 )}
}
}
\usage{
dt_prescripcions_test
}
\description{
Bases de dades de medicaments prescrits
}
\keyword{datasets}
