% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_variables.R
\docType{data}
\name{dt_variables}
\alias{dt_variables}
\title{Prices of 50,000 round cut diamonds}
\format{
A data frame with 50 rows and 4 variables
\describe{
\item{idp}{idp (1--5)}
\item{dat}{dat (20040101--20080101)}
\item{val}{val (0.12--9.25)}
\item{cod}{cod (GLICADA)}
}
}
\usage{
dt_variables
}
\description{
Prices of 50,000 round cut diamonds
}
\keyword{datasets}
