% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_mostra_fitxers.R
\name{generar_mostra_fitxers}
\alias{generar_mostra_fitxers}
\title{Genera mostres.}
\usage{
generar_mostra_fitxers(
  directori = "dades/SIDIAP",
  fitxer_poblacio = "METPLUS_entregable_poblacio_20181126_190346.rds",
  mida_mostra = 10000,
  prefix = "test",
  directori_test = "mostra",
  idp = "idp"
)
}
\arguments{
\item{directori}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{fitxer_poblacio}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{mida_mostra}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{prefix}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{directori_test}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}
}
\value{
La mostra
}
\description{
Llegir tots els fitxers RDS dins d'un directori
i generar una mostra aleatoria i salvar-lo en un directori "mostra"
}
\examples{
u=rnorm(1000,100,12)

}
