% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generar_taula_plana.R
\name{Generar_taula_plana}
\alias{Generar_taula_plana}
\title{Generar la taula plana}
\usage{
Generar_taula_plana(
  dt = "dt_index",
  cataleg = "dt_cataleg",
  parametres = "dt_parametres",
  ...
)
}
\arguments{
\item{dt}{Data.índex}

\item{cataleg}{Catàleg}

\item{parametres}{Paràmetres}

\item{...}{altres funcions}
}
\value{
Retorna una taula plana a partir dels parametres dels agregadors
}
\description{
Generar la taula plana, a partir dels parametres dels agregadors
}
\examples{
#bd data Index.
idp=c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30)
dtindex=c(20220101,20220101,20220101,20220101,20220101,20220101,20220101,20220101,20220101,20220101,
20220101,20220101,20220101,20220101,20220101,20220101,20220101,20220101,20220101,20220101,
20220101,20220101,20220101,20220101,20220101,20220101,20220101,20220101,20220101,20220101)
dt_index<-data.frame(idp=idp,dtindex=dtindex)
dt_index<-tibble::as_tibble(dt_index)
dt_index$dtindex<-as.character(dt_index$dtindex)

#Parametres:
fitxer=c("dt_diagnostics","dt_facturacio","dt_prescripcions","dt_variables")
domini=c("diagnostics","farmacs_facturats","farmacs_prescrits","variables")
Finestra1=c(-Inf,-Inf,-Inf,-Inf)
Finestra2=c(0,0,0,0)
camp=c("agr","agr","agr","cod")
funcio=c("first","first","first","last")
prefix =c("DG.","FF.","FP.","Valor.")
dt_parametres<-data.frame(cbind(fitxer,domini,Finestra1,Finestra2,camp,prefix,funcio))
dt_parametres


Taula_plana<-Generar_taula_plana(
dt=dt_index,
cataleg=cataleg_test,
parametres=dt_parametres)

Taula_plana

Taula_plana2<-Generar_taula_plana(
dt=dt_index,
cataleg=dt_cataleg_test,
parametres=dt_parametres)

Taula_plana2
}
