% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_ggplot_univariant.R
\name{MAP_ggplot_univariant}
\alias{MAP_ggplot_univariant}
\title{Dibuixa mapa temporal}
\usage{
MAP_ggplot_univariant(
  dades = dt,
  datainicial = "data",
  datafinal = "datafi",
  id = "idp_temp",
  Nmostra = 10,
  add_point = NA,
  add_final = NA,
  set_seed = 123
)
}
\arguments{
\item{dades}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{datainicial}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{datafinal}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{id}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{Nmostra}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{add_point}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{add_final}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{set_seed}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}
}
\value{
Dibuixa mapa temporal univariant per verificar solapaments
}
\description{
Dibuixa mapa temporal univariant per verificar solapaments
}
\examples{
u=rnorm(1000,100,12)

}
