% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_punts_ggplot.R
\name{MAP_punts_ggplot}
\alias{MAP_punts_ggplot}
\title{Grafic Analitiques}
\usage{
MAP_punts_ggplot(
  dt = mostra50,
  id = "idp",
  datainicial = "dat",
  val = "val",
  grup_color = "agr",
  Nmostra = Inf,
  llavor = 123,
  finestraX = c(-Inf, +Inf),
  id_AGG = F
)
}
\arguments{
\item{dt}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{id}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{datainicial}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{val}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{grup_color}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{Nmostra}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{llavor}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{finestraX}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{id_AGG}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}
}
\value{
Dibuixa punts grafic analitiques
}
\description{
Grafic Analitiques
}
\examples{
u=rnorm(1000,100,12)

}
