% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_valor_ggplot.R
\name{MAP_valor_ggplot}
\alias{MAP_valor_ggplot}
\title{Grafic Analitiques valors}
\usage{
MAP_valor_ggplot(
  dt = mostra50,
  id = "idp",
  datainicial = "dat",
  val = "val",
  grup_color = "agr",
  Nmostra = 1,
  finestraX = c(-Inf, Inf),
  llavor = 123,
  title = "Evolució de valors"
)
}
\arguments{
\item{dt}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{id}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{datainicial}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{val}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{grup_color}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{Nmostra}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{finestraX}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{llavor}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{title}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}
}
\value{
Dibuixa punts-valors grafic analitiques
}
\description{
Grafic Analitiques valors
}
\examples{
u=rnorm(1000,100,12)

}
