% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normlHist.R
\name{normalHist}
\alias{normalHist}
\title{Dibuixa l'histograma d'una variable superposant-ne la densitat normal ajustada}
\usage{
normalHist(x, dens = FALSE, ...)
}
\arguments{
\item{x}{vector de dades l'histograma del qual es calcularà}

\item{dens}{valor lògic: TRUE=Superposar estimador de nucli de la densitat}

\item{...}{altres funcions}
}
\value{
histograma amb la densitat normal superposada
}
\description{
Funció que dibuixa l'histograma d'una variable x superposant la densitat normal
ajustada. Si l'usuari ho desitja, pot superposar també un estimador de nucli de la densitat
}
\examples{
u=rnorm(1000,100,12)
normalHist(u);
normalHist(u,dens=TRUE)
normalHist(u,dens=TRUE,col="lightcyan")

}
