% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Platan.R
\docType{package}
\name{Platan}
\alias{Platan}
\title{Platan Package}
\description{
Platan és un Package que conté les funcions necessàries,
per aplanar una base de dades per la investigació mèdica
}
\author{
Ray Puig i Jordi Real\email{-raypuig@gmail.com-}
}
