% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_conductor.R
\name{read_conductor}
\alias{read_conductor}
\title{read_conductor}
\usage{
read_conductor(fitxer, ...)
}
\arguments{
\item{fitxer}{fitxer que el convertirem a tibble.}
}
\value{
Tibble
}
\description{
Read conductor file different formats txt o rds o xls xlsx o data_frame tibble
}
\examples{
Read conductor file different formats txt o rds o xls xlsx o data_frame tibble

}
