% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afegir_dataindex.R
\name{afegir_dataindex}
\alias{afegir_dataindex}
\title{Afegeix data-index.}
\usage{
afegir_dataindex(dt_historic, bd.dindex = "20161231")
}
\arguments{
\item{dt_historic}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{bd.dindex}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}
}
\value{
La data índex
}
\description{
Afegeix data-index Dinamica o / Constant si no existeix
}
\examples{
u=rnorm(1000,100,12)

}
