% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_analitiques.R
\name{agregar_analitiques}
\alias{agregar_analitiques}
\title{Agregar analitiques.}
\usage{
agregar_analitiques(
  dt = "ANALITIQUES",
  bd.dindex = "20161231",
  finestra.dies = c(-Inf, Inf),
  sufix = c(".valor", ".dies"),
  fun = "last",
  camp = "cod"
)
}
\arguments{
\item{dt}{La base de dades a on tenim :idp(identificacio Pacient),dat(data del dia de l'analítica:20080101[1-1-2008],val(valor de l'analítica),cod(ex:GLICADA)).}

\item{bd.dindex}{Data a on fem el tall.}

\item{finestra.dies}{Finestra d'observacio a partir de bd.index.}

\item{sufix}{Valor del codi, o dies que han passat.}

\item{fun}{Per defecte Last, la data mes proxima a la bd.index.}

\item{camp}{El camp a on fem l'agreggacio.}
}
\value{
Taula agregada analitiques
}
\description{
Retorna tibble amb els valors analitiques en temps previ en dies.
}
\examples{
dtagr_variables<-agregar_analitiques(dt=dt_variables_test,
bd.dindex=20220101,finestra.dies = c(-Inf,0))
dtagr_variables
}
