% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_facturacio.R
\name{agregar_facturacio}
\alias{agregar_facturacio}
\title{Agregar facturacio.}
\usage{
agregar_facturacio(
  dt = "dt_facturacio",
  finestra.dies = c(-90, 0),
  dt.agregadors = dt.agregadors,
  bd.dindex = "20161231",
  prefix = "FD.",
  camp_agregador = "agr",
  agregar_data = F,
  acumular = NULL,
  cataleg_mana = F
)
}
\arguments{
\item{dt}{Base de dades de Fármac Facturats (idp,cod-A10BB01-,dat-200801,env).}

\item{finestra.dies}{Finestra de dies a partir de la data.index.}

\item{dt.agregadors}{Catàleg, a on tenim els agregadors a prtir del codi.}

\item{bd.dindex}{Data on comencem a contar els dies}

\item{prefix}{Prefix dels agregadors, normalment FF.}

\item{camp_agregador}{Camp agregador.}

\item{agregar_data}{Ens diu la data del fàrmac dins la finestra.}

\item{acumular}{Null.}

\item{cataleg_mana}{Catàleg}
}
\value{
Taula agregada facturacio
}
\description{
Retorna tibble amb la suma d'envasos o data primera dispensació dins d'una finestra de temps per idp-dataindex
}
\examples{

dtagr_facturacio<-agregar_facturacio(
dt=dt_facturacio_test,
bd.dindex=20220101,
finestra.dies=c(-Inf,0),
dt.agregadors=cataleg_test,
prefix="FF.",
camp_agregador="agr",
agregar_data=TRUE,
cataleg_mana=TRUE,
acumular=NULL)
dtagr_facturacio
}
