% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_facturacio.R
\name{agregar_facturacio}
\alias{agregar_facturacio}
\title{Agregar facturacio.}
\usage{
agregar_facturacio(
  dt = PRESCRIPCIONS,
  finestra.dies = c(-90, 0),
  dt.agregadors = CATALEG,
  bd.dindex = "20161231",
  prefix = "FD.",
  camp_agregador = "agr",
  agregar_data = F,
  acumular = NULL,
  cataleg_mana = F
)
}
\arguments{
\item{dt}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{finestra.dies}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{dt.agregadors}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{bd.dindex}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{prefix}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{camp_agregador}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{agregar_data}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{acumular}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{cataleg_mana}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}
}
\value{
Taula agregada facturacio
}
\description{
Retorna tibble (data.table) amb la suma d'envasos o data primera dispensació dins d'una finestra de temps per idp-dataindex
}
\examples{
idp=rep(1:5,each=5)
dat=rep(c(200801,200702,200603,200504,200405),times=5)
cod=rep(c("A10BB01","A10BD01","A10BD04","A10BA02","J01DD07"),times=5)
env=rep(1:5,each=5)
dt_facturacio<-data.frame(idp=idp,cod=cod,dat=dat,env=env)

domini="farmacs_facturats"
cod=c("A10BB01","A10BD01","A10BD04","A10BA02","J01DD07")
agr_Farmac=c("Sulfonilureas","Biguanidas","Tiazolidinadiones","Biguanidas","Antibioticos")
cataleg<-data.frame(domini=domini,cod=cod,agr_Farmac=agr_Farmac)

dtagr_facturacio<-agregar_facturacio(dt=dt_facturacio,bd.dindex=20080120,finestra.dies=c(-90,0),
 dt.agregadors=select(dt_cataleg,cod,agr=agr_Farmac),
 prefix="FF.",
 camp_agregador="agr",
 agregar_data=T,
 cataleg_mana = T)
}
