% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_facturacio.R
\name{agregar_facturacio}
\alias{agregar_facturacio}
\title{Agregar facturacio.}
\usage{
agregar_facturacio(
  dt = "dt_facturacio",
  finestra.dies = c(-90, 0),
  dt.agregadors = dt.agregadors,
  bd.dindex = "20161231",
  prefix = "FD.",
  camp_agregador = "agr",
  agregar_data = F,
  acumular = NULL,
  cataleg_mana = F
)
}
\arguments{
\item{dt}{Base de dades de Fármac Facturats (idp,cod-A10BB01-,dat-200801,env).}

\item{finestra.dies}{Finestra de dies a partir de la data.index.}

\item{dt.agregadors}{Catàleg, a on tenim els agregadors a prtir del codi.}

\item{bd.dindex}{Data on comencem a contar els dies}

\item{prefix}{Prefix dels agregadors, normalment FF.}

\item{camp_agregador}{Camp agregador.}

\item{agregar_data}{Ens diu la data del fàrmac dins la finestra.}

\item{acumular}{Null.}

\item{cataleg_mana}{Catàleg}
}
\value{
Taula agregada facturacio
}
\description{
Retorna tibble (data.table) amb la suma d'envasos o data primera dispensació dins d'una finestra de temps per idp-dataindex
}
\examples{
idp=rep(1:5,each=5)
dat=rep(c(200801,200801,200801,200801,200802),times=5)
cod=rep(c("A10BB01","A10BD01","A10BD04","A10BA02","J01DD07"),times=5)
env=rep(1:5,each=5)
dt_facturacio<-data.frame(idp=idp,cod=cod,dat=dat,env=env)

domini="farmacs_facturats"
cod=c("A10BB01","A10BD01","A10BD04","A10BA02","J01DD07")
agr_Farmac=c("Sulfonilureas","Biguanidas","Tiazolidinadiones","Biguanidas","Antibioticos")
dt_cataleg<-data.frame(domini=domini,cod=cod,agr_Farmac=agr_Farmac)

dt.agregadors=dplyr::select(dt_cataleg,cod,agr=agr_Farmac)

dtagr_facturacio<-agregar_facturacio(
dt=dt_facturacio,
bd.dindex=20080120,
finestra.dies=c(-90,0),
dt.agregadors=dt.agregadors,
prefix="FF.",
camp_agregador="agr",
agregar_data=TRUE,
cataleg_mana=TRUE,
acumular=NULL)
dtagr_facturacio
}
