% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_prescripcions.R
\name{agregar_prescripcions}
\alias{agregar_prescripcions}
\title{Agregar prescripcions.}
\usage{
agregar_prescripcions(
  dt = PRESCRIPCIONS,
  bd.dindex = 20161231,
  dt.agregadors = CATALEG,
  prefix = "FP.",
  finestra.dies = c(0, 0),
  camp_agregador = "agr",
  agregar_data = F,
  acumular = NULL,
  cataleg_mana = F
)
}
\arguments{
\item{dt}{Base de dades de Fármac Prescrips (idp,cod-A10BB01-,dat-200801).}

\item{bd.dindex}{xData on comencem a contar els dies.}

\item{dt.agregadors}{Catàleg, a on tenim els agregadors a prtir del codi.}

\item{prefix}{Prefix dels agregadors, normalment FP.}

\item{finestra.dies}{Finestra de dies a partir de la data.index.}

\item{camp_agregador}{Camp agregador.}

\item{agregar_data}{Ens diu la data del fàrmac dins la finestra.}

\item{acumular}{Null.}

\item{cataleg_mana}{Catàleg}
}
\value{
Taula agregada prescripcions
}
\description{
Retorna tibble (data.table) amb el temps de prescripció en una finestra o primera data per idp-dataindex / primera data
}
\examples{
idp=rep(1:5,each=5)
dat=rep(c(20080115,20080115,20080115,20080115,20080215),times=5)
dbaixa=rep(c(20080215,20080215,20080215,20080215,20080315),times=5)
cod=rep(c("A10BB01","A10BD01","A10BD04","A10BA02","J01DD07"),times=5)
dt_facturacio<-data.frame(idp=idp,cod=cod,dat=dat,dbaixa=dbaixa)

domini="farmacs_prescrits"
cod=c("A10BB01","A10BD01","A10BD04","A10BA02","J01DD07")
agr_Farmac=c("Sulfonilureas","Biguanidas","Tiazolidinadiones","Biguanidas","Antibioticos")
dt_cataleg<-data.frame(domini=domini,cod=cod,agr_Farmac=agr_Farmac)

dtagr_facturacio<-agregar_prescripcions(
dt=dt_facturacio,
bd.dindex=20080120,
finestra.dies=c(-90,0),
dt.agregadors=select(dt_cataleg,cod,agr=agr_Farmac),
prefix="FP.",
camp_agregador="agr",
agregar_data=T,
cataleg_mana = T,
acumular=NULL)
}
