% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_prescripcions.R
\name{agregar_prescripcions}
\alias{agregar_prescripcions}
\title{Agregar prescripcions.}
\usage{
agregar_prescripcions(
  dt = "PRESCRIPCIONS",
  bd.dindex = 20161231,
  dt.agregadors = "CATALEG",
  prefix = "FP.",
  finestra.dies = c(0, 0),
  camp_agregador = "agr",
  agregar_data = F,
  acumular = NULL,
  cataleg_mana = F
)
}
\arguments{
\item{dt}{Base de dades de Fármac Prescrips (idp,cod-A10BB01-,dat-200801).}

\item{bd.dindex}{Data on comencem a contar els dies.}

\item{dt.agregadors}{Catàleg, a on tenim els agregadors a prtir del codi.}

\item{prefix}{Prefix dels agregadors, normalment FP.}

\item{finestra.dies}{Finestra de dies a partir de la data.index.}

\item{camp_agregador}{Camp agregador.}

\item{agregar_data}{Ens diu la data del fàrmac dins la finestra.}

\item{acumular}{Null.}

\item{cataleg_mana}{Catàleg}
}
\value{
Taula agregada prescripcions
}
\description{
Retorna tibble (data.table) amb el temps de prescripció en una finestra o primera data per idp-dataindex / primera data
}
\examples{

dtagr_prescripcions<-agregar_prescripcions(
dt=dt_prescripcions,
bd.dindex=20220101,
finestra.dies=c(-Inf,0),
dt.agregadors=cataleg,
prefix="FP.",
camp_agregador="agr",
agregar_data=TRUE,
cataleg_mana =TRUE,
acumular=NULL)
dtagr_prescripcions
}
