% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_problemes.R
\name{agregar_problemes}
\alias{agregar_problemes}
\title{Agregar problemes de salut.}
\usage{
agregar_problemes(
  dt = "PROBLEMES",
  bd.dindex = "20161231",
  dt.agregadors = "CATALEG",
  finestra.dies = c(-Inf, 0),
  prefix = "DG.",
  camp_agregador = "agr",
  keep.code = F,
  cataleg_mana = F
)
}
\arguments{
\item{dt}{Base de dades dels Diàgnostics (idp,cod-E11.5-,dat-20150309).}

\item{bd.dindex}{Data on comencem a contar els dies.}

\item{dt.agregadors}{Catàleg, a on tenim els agregadors a prtir del codi.}

\item{finestra.dies}{Finestra de dies a partir de la data.index.}

\item{prefix}{Prefix dels agregadors, normalment DG.}

\item{camp_agregador}{Camp agregador.}

\item{keep.code}{Guarda els codis.}

\item{cataleg_mana}{Catàleg.}
}
\value{
Taula agregada de problemes
}
\description{
Afegeix data-index Dinamica o / Constant si no existeix
}
\examples{

dtagr_diagnostics<-agregar_problemes(dplyr::select(dt_diagnostics_test,idp,cod,dat),
bd.dindex = 20220101,
dt.agregadors=cataleg_test,
finestra.dies=c(-Inf,0),prefix = "DG.",
cataleg_mana=TRUE)

dtagr_diagnostics

}
