% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_solapaments_gaps.R
\name{agregar_solapaments_gaps}
\alias{agregar_solapaments_gaps}
\title{agregar_solapaments_gaps}
\usage{
agregar_solapaments_gaps(
  dt = "dades",
  id = "idp",
  datainici = "data",
  datafinal = "datafi",
  gap = 5,
  sel = F
)
}
\arguments{
\item{dt}{dt}

\item{id}{id}

\item{datainici}{datainici}

\item{datafinal}{datafinal}

\item{gap}{gap}

\item{sel}{sel}
}
\description{
agregar_solapaments_gaps
}
