% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_solapaments_gaps.R
\name{agregar_solapaments_gaps}
\alias{agregar_solapaments_gaps}
\title{Agrega solapaments amb gaps}
\usage{
agregar_solapaments_gaps(
  dt = dades,
  id = "idp",
  datainici = "data",
  datafinal = "datafi",
  gap = 5,
  sel = FALSE
)
}
\arguments{
\item{dt}{Base de dades de farmacs}

\item{id}{Identificador}

\item{datainici}{Data Inici Farmac}

\item{datafinal}{Data Final Farmac}

\item{gap}{GAPS}

\item{sel}{Seleccionem}
}
\value{
Retorna dades amb els gaps (discontinuitat petita)
}
\description{
Historic de farmacs.Elimina solapaments.
}
\examples{
idp=rep(1:5,each=5)
dat=rep(c(20080115,20080115,20080115,20080115,20080215),times=5)
dbaixa=rep(c(20080215,20080215,20080215,20080215,20080315),times=5)
cod=rep(c("A10BB01","A10BD01","A10BD04","A10BA02","J01DD07"),times=5)
dt_prescripcions<-data.frame(idp=idp,cod=cod,dat=dat,dbaixa=dbaixa)
gaps<-agregar_solapaments_gaps(dt=dt_prescripcions,
id="idp",
datainici="dat",
datafinal="dbaixa",
gap=2,sel=FALSE)

gaps
}
