% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_diagnostics.R
\docType{data}
\name{dt_diagnostics}
\alias{dt_diagnostics}
\title{Bases de dades de diagnostics}
\format{
Una data frame amb 25 files i 3 variables
\describe{
\item{idp}{idp (1--5)}
\item{cod}{codi (E11   I10   I25 I50.9 )}
\item{dat}{data (20080115--20080215)}
}
}
\usage{
dt_diagnostics
}
\description{
Bases de dades de diagnostics
}
\keyword{datasets}
