% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_facturacio.R
\docType{data}
\name{dt_facturacio}
\alias{dt_facturacio}
\title{Bases de dades de medicaments facturats}
\format{
Una data frame amb 25 files i 4 variables
\describe{
\item{idp}{idp (1--5)}
\item{cod}{codi (A10BA02 A10BB01 A10BD01 A10BD04 J01DD07 )}
\item{dat}{data (200801--200802)}
\item{env}{envasos (1--5)}
}
}
\usage{
dt_facturacio
}
\description{
Bases de dades de medicaments facturats
}
\keyword{datasets}
