% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_variables.R
\docType{data}
\name{dt_variables}
\alias{dt_variables}
\title{Base de dades analitiques i cliniques}
\format{
Una data frame amb 25 files i 4 variables
\describe{
\item{idp}{idp (1--5)}
\item{dat}{data (20040101--20080101)}
\item{val}{valor (0.12--9.25)}
\item{cod}{codi (GLICADA)}
}
}
\usage{
dt_variables
}
\description{
Base de dades analitiques i cliniques
}
\keyword{datasets}
