% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_mostra_fitxers.R
\name{generar_mostra_fitxers}
\alias{generar_mostra_fitxers}
\title{Genera mostres.}
\usage{
generar_mostra_fitxers(
  directori = "data",
  fitxer_poblacio = "METPLUS_entregable_poblacio_20181126_190346.rds",
  mida_mostra = 10000,
  prefix = "test",
  directori_test = "mostra",
  idp = "idp",
  llavor = "1234"
)
}
\arguments{
\item{directori}{Directori a on tenim les DADES ORIGINALS}

\item{fitxer_poblacio}{Nom del fitxer de Poblacio}

\item{mida_mostra}{Mida de la mostra}

\item{prefix}{Prefix de la mostra}

\item{directori_test}{Directori Test per la nostra MOSTRA}

\item{idp}{idp}

\item{llavor}{llavor aleatoria}
}
\value{
La mostra
}
\description{
Llegir tots els fitxers RDS dins d'un directori generar una mostra aleatoria i salvar-lo en un directori mostra
}
