% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_mostra_fitxers.R
\name{generar_mostra_fitxers}
\alias{generar_mostra_fitxers}
\title{Genera mostres.}
\usage{
generar_mostra_fitxers(
  directori = "data",
  fitxer_poblacio = "METPLUS_entregable_poblacio_20181126_190346.rds",
  mida_mostra = 10000,
  prefix = "test",
  directori_test = "mostra",
  idp = "idp"
)
}
\arguments{
\item{directori}{Directori a on tenim les dades}

\item{fitxer_poblacio}{Nom del fitxer de Poblacio}

\item{mida_mostra}{Mida de la mostra}

\item{prefix}{Prefix de la mostra}

\item{directori_test}{Directori Test}
}
\value{
La mostra
}
\description{
Llegir tots els fitxers RDS dins d'un directorii generar una mostra aleatoria i salvar-lo en un directori mostra
}
\examples{
idp=c(1,2,3,4,5)
sex=c("H","H","D","D","H")
edat=c(23,67,90,16,100)
dt_poblacio<-data.frame(idp=idp,sex=sex,edat=edat)
saveRDS(dt_poblacio, file=here::here("data","dt_poblacio.rds"))

idp=rep(1:5,each=5)
dat=rep(c(200801,200801,200801,200801,200802),times=5)
cod=rep(c("A10BB01","A10BD01","A10BD04","A10BA02","J01DD07"),times=5)
env=rep(1:5,each=5)
dt_facturacio<-data.frame(idp=idp,cod=cod,dat=dat,env=env)
saveRDS(dt_facturacio, file=here::here("data","dt_facturacio.rds"))

idp=rep(1:5,each=5)
dat=rep(c(20080115,20080115,20080115,20080115,20080215),times=5)
dbaixa=rep(c(20080215,20080215,20080215,20080215,20080315),times=5)
cod=rep(c("A10BB01","A10BD01","A10BD04","A10BA02","J01DD07"),times=5)
dt_prescripcions<-data.frame(idp=idp,cod=cod,dat=dat,dbaixa=dbaixa)
saveRDS(dt_prescripcions, file=here::here("data","dt_prescripcions.rds"))

idp=rep(1:5,each=5)
dat=rep(c(20080115,20080115,20080115,20080115,20080215),times=5)
cod=rep(c("E11","E11","I25","I50.9","I10"),times=5)
dt_diagnostics<-data.frame(idp=idp,cod=cod,dat=dat)
saveRDS(dt_diagnostics, file=here::here("data","dt_diagnostics.rds"))

idp=rep(1:5,each=5)
dat=rep(c(20080101,20070101,20060101,20050101,20040101),times=5)
val=round(rnorm(50,5,1.9),digits=2)
cod="GLICADA"
dt_variables<-data.frame(idp=idp,dat=dat,val=val,cod=cod)
saveRDS(dt_variables, file=here::here("data","dt_variables.rds"))

MOSTRA<-generar_mostra_fitxers(directori="data",
fitxer_poblacio="dt_poblacio.rds",
mida_mostra=3,prefix="test_",
directori_test="mostra",
idp="idp")

test_dt_diagnostics<-readRDS(here::here("data/mostra","test_dt_diagnostics.rds"))
test_dt_diagnostics
}
