% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generar_taula_plana.R
\name{Generar_taula_plana}
\alias{Generar_taula_plana}
\title{Generar la taula plana, a partir dels parametres dels agregadors}
\usage{
Generar_taula_plana(
  dt = dt_index,
  cataleg = dt_cataleg,
  parametres = dt_parametres,
  ...
)
}
\arguments{
\item{dt}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{cataleg}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

\item{parametres}{xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}
}
\value{
Retorna una taula plana a partir dels parametres dels agregadors
}
\description{
Taula plana
}
\examples{
u=rnorm(1000,100,12)

}
