% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Platan.R
\docType{package}
\name{Platan}
\alias{Platan}
\title{Platan Package}
\description{
Platan es un Package que conte les funcions necessaries,
per aplanar una base de dades per la investigació medica
}
\author{
Jordi Real i Ray Puig\email{-raypuig@gmail.com-}
}
