% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_conductor.R
\name{read_conductor}
\alias{read_conductor}
\title{read_conductor}
\usage{
read_conductor(fitxer, ...)
}
\arguments{
\item{fitxer}{fitxer que el convertirem a tibble.}

\item{...}{altres funcions}
}
\value{
una taula tibble
}
\description{
Llegir el fitxer conductor de diferents formats txt o rds o xls xlsx o data_frame tibble
}
